backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
argon2-cffi
